Added a non-functional devlog input system
Won’t be functional until we get a server unless there is a work around
for forms. Will also need to implement a feature where this is only
displayed for devs. I am thinking of a dev tag on usernames then a
simple if statement. Will have to see if this is possible with
Javascript although I may have to create a separate page for this.